DBM: fixed potential overflow in FLOPS calculation

+++ src/dbm/dbm_multiply.c
@@ -279,7 +279,7 @@ static void multiply_packs(const bool transa, const bool transb,
 
             // Count flops.
             dbm_library_counter_increment(m, n, k);
-            const int task_flops = 2 * m * n * k;
+            const int64_t task_flops = 2LL * m * n * k;
             flop_sum += task_flops;
             if (task_flops == 0) {
               continue;
